7_7_app/requirements.txt
# getRNNData requirements.txt 

numpy==1.18.4
torch==1.5.0
argparse==1.4.0
pandas==1.0.3
PyYAML==5.3
# torch.utils.data
# copy
# pickle
# pdb
